Move AppRelease to library
